New and Improved MC, with detector information
Currently this works for the muon part of detection.  Detailed
information on which triggers fire and where in the detector muons go
and decay is given.  Can force the muon to decay and give signal, or can
set the decay rate lower.  Still need to add in electron decays and then
have it output to a TTree.  Runs very fast; 10 million events per
minute.